Organize tests into unit, integration and e2e folders
